new module added for reduce module lenght
